Create an Apache Thrift language binding for Dart (dartlang.org). https://issues.apache.org/jira/browse/THRIFT-3299
Protocol and transport unit tests.
